Make aspects first-class plugins
